requirements.txt
Flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
six==1.16.0
Werkzeug==2.2.3
aiohttp
bs4
python-dotenv